backend/pyproject.toml
[project]
name = "neofish-backend"
version = "0.1.0"
description = "Neofish - 간결하고 범용적인 집단지성 엔진, 무엇이든 예측"
requires-python = ">=3.11"
license = { text = "AGPL-3.0" }
authors = [
    { name = "Neofish Team" }
]

dependencies = [
    # 핵심 프레임워크
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",
    # LLM 관련
    "openai>=1.0.0",
    "neo4j>=5.14.0",
    # OASIS 소셜 미디어 시뮬레이션
    "camel-oasis==0.2.5",
    "camel-ai==0.2.78",
    # 파일 처리
    "PyMuPDF>=1.24.0",
    # 인코딩 감지(UTF-8이 아닌 텍스트 파일 지원)
    "charset-normalizer>=3.0.0",
    "chardet>=5.0.0",
    # 유틸리티 라이브러리
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "orjson>=3.10.0",
    "google-genai>=1.63.0",
    "pydub>=0.25.1",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pipreqs>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
